project-service/pyproject.toml
[tool.poetry]
name = "unicollab"
version = "0.1.0"
description = "A collaboration platform for university students to manage projects"
authors = ["ayatsider <22011430@students.hebron.edu>"]
readme = "../README.md"
packages = [
    {include = "app"},
    {include = "cli"}
]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = ">=0.129.0,<0.130.0"
uvicorn = ">=0.40.0,<0.41.0"
requests = ">=2.32.5,<3.0.0"
python-dotenv = ">=1.2.1,<2.0.0"
pydantic = ">=2.12.5,<3.0.0"
pymongo = ">=4.16.0,<5.0.0"
httpx = ">=0.28.1,<0.29.0"
pyjwt = "^2.11.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-cov = ">=7.0.0,<8.0.0"
black = ">=26.1.0,<27.0.0"
isort = ">=7.0.0,<8.0.0"
flake8 = ">=7.3.0,<8.0.0"
coverage = ">=7.13.4,<8.0.0"

[tool.poetry.scripts]
unicollab-cli = "cli.project_cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
